exercise 4.21

+++ part4/controllers/blogs.js
@@ -17,7 +17,6 @@ blogsRouter.post('/', async (request, response, next) => {
   }
 
   try {
-    console.log('token: ', request.token)
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
     if (!request.token || !decodedToken.id) {
       return response.status(401).json({ error: 'token missing or invalid' })
@@ -57,7 +56,22 @@ blogsRouter.get('/:id', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+    if (!blog) {
+      return response.status(401).json({ error: 'blog missing' })
+    }
+
+    if (decodedToken.id.toString() !== blog.user.toString()) {
+      return response.status(401).end()
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
+
     response.status(204).end()
   } catch (exception) {
     next(exception)
